rename trig.rs -> cossin.rs
