+++ dsp/src/cossin.rs
